backend/requirements.txt
python-multipart
fastapi
uvicorn
torch
torchvision
librosa
numpy
pillow

